Получить всех менти одного пользователя. Fixed.

+++ src/main/java/school/faang/user_service/controller/mentorship/MentorshipController.java
@@ -1,5 +1,6 @@
 package school.faang.user_service.controller.mentorship;
 
+import jakarta.validation.constraints.Positive;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -14,7 +15,10 @@ public class MentorshipController {
     private final MentorshipService mentorshipService;
 
     @DeleteMapping("/mentee")
-    public void deleteMentee(@RequestParam long mentorId, @RequestParam long menteeId) {
+    public void deleteMentee(@RequestParam @Positive long mentorId, @RequestParam @Positive long menteeId) {
+        if (mentorId == menteeId) {
+            throw new UnsupportedOperationException("You cannot delete the mentee-user who is the same as the mentor-user.");
+        }
         mentorshipService.deleteMentee(mentorId, menteeId);
     }
 }
